Added explicit casts in read methods of FakeMemory.

+++ src/fake_memory.cc
@@ -145,7 +145,7 @@ uint8_t FakeMemory::read_byte(uint64_t address) { return read(address, 1)[0]; }
  */
 uint16_t FakeMemory::read_half_word(uint64_t address) {
     std::vector<uint8_t> data = read(address, 2);
-    uint16_t half_word = (data[1] << 8) | data[0];
+    uint16_t half_word = (((uint16_t)data[1]) << 8) | data[0];
     return half_word;
 }
 
@@ -156,7 +156,8 @@ uint16_t FakeMemory::read_half_word(uint64_t address) {
  */
 uint32_t FakeMemory::read_word(uint64_t address) {
     std::vector<uint8_t> data = read(address, 4);
-    uint32_t word = (data[3] << 24) | (data[2] << 16) | (data[1] << 8) | data[0];
+    uint32_t word = (((uint32_t)data[3]) << 24) | (((uint32_t)data[2]) << 16) |
+                    (((uint32_t)data[1]) << 8) | data[0];
     return word;
 }
 
@@ -167,9 +168,10 @@ uint32_t FakeMemory::read_word(uint64_t address) {
  */
 uint64_t FakeMemory::read_double_word(uint64_t address) {
     std::vector<uint8_t> data = read(address, 4);
-    uint64_t double_word = (data[7] << 56) | (data[6] << 48) | (data[5] << 40) |
-                           (data[4] << 32) | (data[3] << 24) | (data[2] << 16) |
-                           (data[1] << 8) | data[0];
+    uint64_t double_word = (((uint64_t)data[7]) << 56) | (((uint64_t)data[6]) << 48) |
+                           (((uint64_t)data[5]) << 40) | (((uint64_t)data[4]) << 32) |
+                           (((uint64_t)data[3]) << 24) | (((uint64_t)data[2]) << 16) |
+                           (((uint64_t)data[1]) << 8) | data[0];
     return double_word;
 }
